+++ app/src/main/java/scot/raven/titanpad/settings/ui/setup/TutorialActivity.kt
@@ -1,0 +1,26 @@
+package scot.raven.titanpad.settings.ui.setup
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.lifecycle.ViewModelProvider
+import scot.raven.titanpad.TitanPad
+import scot.raven.titanpad.core.ui.AppTheme
+import scot.raven.titanpad.settings.ui.SettingsActivity.Companion.CONFIG_ID_EXTRA
+import scot.raven.titanpad.settings.ui.SettingsState
+
+class TutorialActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        setContent {
+            AppTheme {
+                TutorialScreen(
+                    onNavigateBack = {
+                        finish()
+                    }
+                )
+            }
+        }
+    }
+}