+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for backing up and restoring a single configuration to a json file. Can be used to duplicate configs or to share configs with others
 - Add support for auto-enable config on boot. Useful to counter the effects when killing the app in the app switcher. Also useful during boot if Shizuku can auto-start
 - Add support for hardware mouse wheel as an input. This will emulate a mouse with two scroll wheels, both capable of high resolution movements. Options in the menu also include emulating momentum based wheel as well. Wheel would create the scroll effect at the position the hardware mouse is at.
+- Add tutorial screen linking to a couple tutorial videos on YouTube
 
 ### Changed